Problem: I don't have a blog anymore
Solution: simplify previous blog code (manpages/lue), remove categories.
Figure out how to build `hakyll` in 2018. All in all, it posed another problem,
which is Unicode doesn't work on NixOS #1, also cache doesn't work as expected #2.
In regards to the new problems, I'm going to try to run the same code on a
Ubuntu machine and see if the problem persists.